modify tests to test
