+++ sofa-serverless-runtime/arklet-core/src/test/java/com/alipay/sofa/serverless/arklet/core/command/handler/HealthHandlerTest.java
@@ -30,7 +30,7 @@ import org.junit.Test;
 /**
  * @author lunarscave
  */
-public class HealthHandlerTests extends BaseHandlerTest {
+public class HealthHandlerTest extends BaseHandlerTest {
 
     private HealthHandler handler;
