+++ sofa-serverless-runtime/arklet-core/src/test/java/com/alipay/sofa/serverless/arklet/core/command/handler/InstallBizHandlerTest.java
@@ -16,10 +16,9 @@
  */
 package com.alipay.sofa.serverless.arklet.core.command.handler;
 
-import com.alipay.sofa.ark.api.ClientResponse;
 import com.alipay.sofa.serverless.arklet.core.command.builtin.BuiltinCommand;
-import com.alipay.sofa.serverless.arklet.core.command.builtin.handler.UninstallBizHandler;
-import com.alipay.sofa.serverless.arklet.core.command.handler.BaseHandlerTest;
+import com.alipay.sofa.serverless.arklet.core.command.builtin.handler.InstallBizHandler;
+import com.alipay.sofa.serverless.arklet.core.command.builtin.handler.InstallBizHandler.Input;
 import com.alipay.sofa.serverless.arklet.core.command.meta.Output;
 import com.alipay.sofa.serverless.arklet.core.common.exception.CommandValidationException;
 import org.junit.Assert;
@@ -28,51 +27,54 @@ import org.junit.Test;
 
 import static org.mockito.Mockito.when;
 
-public class UninstallBizHandlerTests extends BaseHandlerTest {
+/**
+ * @author mingmen
+ * @date 2023/9/5
+ */
+public class InstallBizHandlerTest extends BaseHandlerTest {
 
-    private UninstallBizHandler handler;
+    private InstallBizHandler handler;
 
     @Before
     public void setupInstallBizHandler() {
-        handler = (UninstallBizHandler) commandService.getHandler(BuiltinCommand.UNINSTALL_BIZ);
+        handler = (InstallBizHandler) commandService.getHandler(BuiltinCommand.INSTALL_BIZ);
     }
 
     @Test
     public void testHandle_Success() throws Throwable {
-        UninstallBizHandler.Input input = new UninstallBizHandler.Input();
-        input.setBizName("testBizName");
-        input.setBizVersion("testBizVersion");
+        Input input = new Input();
+        input.setBizUrl("testUrl");
 
-        when(handler.getOperationService().uninstall(input.getBizName(), input.getBizVersion()))
-            .thenReturn(success);
+        when(handler.getOperationService().install(input.getBizUrl())).thenReturn(success);
 
-        Output<ClientResponse> result = handler.handle(input);
+        Output<InstallBizHandler.InstallBizClientResponse> result = handler.handle(input);
 
-        Assert.assertEquals(success, result.getData());
+        Assert.assertEquals(success.getBizInfos(), result.getData().getBizInfos());
+        Assert.assertEquals(success.getMessage(), result.getData().getMessage());
+        Assert.assertEquals(success.getCode(), result.getData().getCode());
         Assert.assertTrue(result.success());
     }
 
     @Test
     public void testHandle_Failure() throws Throwable {
-        UninstallBizHandler.Input input = new UninstallBizHandler.Input();
-        input.setBizName("testBizName");
-        input.setBizVersion("testBizVersion");
+        Input input = new Input();
+        input.setBizUrl("testUrl");
 
-        when(handler.getOperationService().uninstall(input.getBizName(), input.getBizVersion()))
-            .thenReturn(failed);
+        when(handler.getOperationService().install(input.getBizUrl())).thenReturn(failed);
 
-        Output<ClientResponse> result = handler.handle(input);
+        Output<InstallBizHandler.InstallBizClientResponse> result = handler.handle(input);
 
-        Assert.assertSame(failed, result.getData());
+        Assert.assertEquals(failed.getBizInfos(), result.getData().getBizInfos());
+        Assert.assertEquals(failed.getMessage(), result.getData().getMessage());
+        Assert.assertEquals(failed.getCode(), result.getData().getCode());
         Assert.assertTrue(result.failed());
     }
 
     @Test(expected = CommandValidationException.class)
     public void testValidate_BlankBizName() throws CommandValidationException {
         // 准备测试数据
-        UninstallBizHandler.Input input = new UninstallBizHandler.Input();
+        Input input = new Input();
         input.setBizName("");
-        input.setBizVersion("testBizVersion");
 
         // 执行测试
         handler.validate(input);
@@ -81,8 +83,7 @@ public class UninstallBizHandlerTests extends BaseHandlerTest {
     @Test(expected = CommandValidationException.class)
     public void testValidate_BlankBizVersion() throws CommandValidationException {
         // 准备测试数据
-        UninstallBizHandler.Input input = new UninstallBizHandler.Input();
-        input.setBizName("testBizName");
+        Input input = new Input();
         input.setBizVersion("");
 
         // 执行测试
@@ -92,9 +93,19 @@ public class UninstallBizHandlerTests extends BaseHandlerTest {
     @Test(expected = CommandValidationException.class)
     public void testValidate_BlankRequestId() throws CommandValidationException {
         // 执行测试
-        UninstallBizHandler.Input input = new UninstallBizHandler.Input();
+        Input input = new Input();
         input.setAsync(true);
         input.setRequestId("");
+
+        // 执行测试
+        handler.validate(input);
+    }
+
+    @Test(expected = CommandValidationException.class)
+    public void testValidate_BlankBizUrl() throws CommandValidationException {
+        // 准备测试数据
+        Input input = new Input();
+        input.setBizUrl("");
 
         // 执行测试
         handler.validate(input);
