+++ sofa-serverless-runtime/arklet-springboot-starter/src/test/java/com/alipay/sofa/serverless/arklet/spring/health/endpoint/ArkHealthCodeEndpointTest.java
@@ -28,7 +28,7 @@ import org.junit.Test;
 /**
  * @author Lunarscave
  */
-public class ArkHealthCodeEndpointTests extends SpringbootBaseTest {
+public class ArkHealthCodeEndpointTest extends SpringbootBaseTest {
     private static ArkHealthCodeEndpoint arkHealthCodeEndpoint;
     private static String                bizName;
     private static String                bizVersion;
