+++ sofa-serverless-runtime/arklet-springboot-starter/src/test/java/com/alipay/sofa/serverless/arklet/spring/health/endpoint/ArkHealthzEndpointTest.java
@@ -31,7 +31,7 @@ import java.util.Map;
 /**
  * @author Lunarscave
  */
-public class ArkHealthzEndpointTests extends SpringbootBaseTest {
+public class ArkHealthzEndpointTest extends SpringbootBaseTest {
 
     private static ArkHealthzEndpoint arkHealthzEndpoint;
     private static String             bizName;
